Module Odoc_search

module Entry : sig ... end

module Html : sig ... end

module Text : sig ... end

This module is used to turn values from odoc model into a more easily searchable format: text! HTML also is text, but it has tags that are to be ignored by a search engine.